+++ src/Instance/InstanceDetails/InstanceAdministrativeView/InstanceAdministrativeView.js
@@ -10,7 +10,8 @@ import {
   Row,
   Col,
   KeyValue,
-  Highlighter
+  Highlighter,
+  NoValue,
 } from '@folio/stripes/components';
 import {
   ViewMetaData,
@@ -26,8 +27,12 @@ import {
   AdministrativeNoteList,
 } from '../../../components';
 
+import {
+  QUERY_INDEXES,
+  SOURCE_VALUES,
+} from '../../../constants';
+
 import StatisticalCodesList from './StatisticalCodesList';
-import { QUERY_INDEXES } from '../../../constants';
 
 const InstanceAdministrativeView = ({
   id,
@@ -65,6 +70,14 @@ const InstanceAdministrativeView = ({
 
     return (instance.statisticalCodeIds || []).map(codeId => statisticalCodesMap[codeId]);
   }, [instance, statisticalCodeTypes, statisticalCodes]);
+
+  const getSourceValue = (source) => {
+    if (!source || source === '-') {
+      return <NoValue />;
+    }
+
+    return SOURCE_VALUES[source] ?? source;
+  };
 
   const queryHRID = queryString.parse(search)?.query;
   const isQueryByHRID = queryString.parse(search)?.qindex === QUERY_INDEXES.INSTANCE_HRID;
@@ -117,7 +130,7 @@ const InstanceAdministrativeView = ({
         <Col xs={3}>
           <KeyValue
             label={<FormattedMessage id="ui-inventory.metadataSource" />}
-            value={checkIfElementIsEmpty(instance.source)}
+            value={getSourceValue(instance.source)}
           />
         </Col>