+++ src/ViewInstance.js
@@ -13,6 +13,7 @@ import {
   IfPermission,
   Pluggable,
   stripesConnect,
+  checkIfUserInMemberTenant,
 } from '@folio/stripes/core';
 import {
   Pane,
@@ -39,6 +40,7 @@ import {
   isUserInConsortiumMode,
 } from './utils';
 import {
+  CONSORTIUM_PREFIX,
   indentifierTypeNames,
   layers,
   REQUEST_OPEN_STATUSES,
@@ -437,6 +439,10 @@ class ViewInstance extends React.Component {
     return null;
   };
 
+  hasCentralTenantPerm = (perm) => {
+    return this.props.centralTenantPermissions.some(({ permissionName }) => permissionName === perm);
+  }
+
   createActionMenuGetter = (instance, data) => ({ onToggle }) => {
     const {
       canUseSingleRecordImport,
@@ -453,14 +459,18 @@ class ViewInstance extends React.Component {
       titleLevelRequestsFeatureEnabled,
     } = this.state;
 
+    const editBibRecordPerm = 'ui-quick-marc.quick-marc-editor.all';
+    const editInstancePerm = 'ui-inventory.instance.edit';
     const isSourceMARC = isMARCSource(instance?.source);
-    const canEditInstance = stripes.hasPerm('ui-inventory.instance.edit');
+    const canEditInstance = stripes.hasPerm(editInstancePerm);
     const canCreateInstance = stripes.hasPerm('ui-inventory.instance.create');
     const canCreateRequest = stripes.hasPerm('ui-requests.create');
     const canMoveItems = stripes.hasPerm('ui-inventory.item.move');
     const canCreateMARCHoldings = stripes.hasPerm('ui-quick-marc.quick-marc-holdings-editor.create');
     const canMoveHoldings = stripes.hasPerm('ui-inventory.holdings.move');
-    const canEditMARCRecord = stripes.hasPerm('ui-quick-marc.quick-marc-editor.all');
+    const canEditMARCRecord = checkIfUserInMemberTenant(stripes) && checkIfSharedInstance(stripes, instance)
+      ? this.hasCentralTenantPerm(editBibRecordPerm)
+      : stripes.hasPerm(editBibRecordPerm);
     const canDeriveMARCRecord = stripes.hasPerm('ui-quick-marc.quick-marc-editor.duplicate');
     const hasReorderPermissions = canCreateRequest || stripes.hasPerm('ui-requests.edit') || stripes.hasPerm('ui-requests.all');
     const canViewMARCSource = stripes.hasPerm('ui-quick-marc.quick-marc-editor.view');
@@ -481,6 +491,10 @@ class ViewInstance extends React.Component {
         onClickHandler(this.context.sendCallout);
       };
     };
+
+    const suppressEditInstanceForMemberTenant = checkIfUserInMemberTenant(stripes)
+      && instance?.source.startsWith(CONSORTIUM_PREFIX)
+      && !this.hasCentralTenantPerm(editInstancePerm);
 
     const showInventoryMenuSection = (
       canEditInstance
@@ -509,7 +523,7 @@ class ViewInstance extends React.Component {
       <>
         {showInventoryMenuSection && (
           <MenuSection label={intl.formatMessage({ id: 'ui-inventory.inventory.label' })} id="inventory-menu-section">
-            {canEditInstance && (
+            {canEditInstance && !suppressEditInstanceForMemberTenant && (
               <Button
                 id="edit-instance"
                 onClick={() => {
@@ -764,6 +778,7 @@ class ViewInstance extends React.Component {
       updateLocation,
       canUseSingleRecordImport,
       intl,
+      isCentralTenantPermissionsLoading,
     } = this.props;
     const ci = makeConnectedInstance(this.props, stripes.logger);
     const instance = ci.instance();
@@ -799,7 +814,7 @@ class ViewInstance extends React.Component {
       },
     ];
 
-    if (!instance) {
+    if (!instance || isCentralTenantPermissionsLoading) {
       return (
         <Pane
           id="pane-instancedetails"
@@ -902,6 +917,7 @@ class ViewInstance extends React.Component {
 
 ViewInstance.propTypes = {
   canUseSingleRecordImport: PropTypes.bool,
+  centralTenantPermissions: PropTypes.arrayOf(PropTypes.object).isRequired,
   selectedInstance:  PropTypes.object,
   goTo: PropTypes.func.isRequired,
   location: PropTypes.shape({
@@ -920,6 +936,7 @@ ViewInstance.propTypes = {
   intl: PropTypes.shape({
     formatMessage: PropTypes.func.isRequired,
   }),
+  isCentralTenantPermissionsLoading: PropTypes.bool.isRequired,
   mutator: PropTypes.shape({
     allInstanceItems: PropTypes.object.isRequired,
     holdings: PropTypes.shape({