+++ CHANGELOG.md
@@ -59,6 +59,9 @@
 * Disable "Save & close" button by default in "Create new Item" window. Fixes UIIN-2492.
 * Inventory 2nd pane Actions menu: Adjust New action. Refs UIIN-2439.
 * Add "Local" or "Shared" to flag MARC bib records. Refs UIIN-2522.
+* Edit MARC bib record on Shared Instance. Refs UIIN-2526.
+* Allow for new source values for shadow Instances. Refs UIIN-2459.
+* Update Node.js to v18 in GitHub Actions. Refs UIIN-2520.
 * Fix issue with Relevance sort option. Refs UIIN-2474.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)